refactored Profile and parts  Database into Database and Analisys | v2 | @drbrx @wlankasper
